pyproject.toml
[tool.poetry]
name = "nuruja"
version = "0.1.0"
description = ""
authors = ["Gichana <gichanamayaka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
pendulum = "^2.1.2"
flask-pydantic = "^0.11.0"
python-dotenv = "^1.0.0"
faker = "^18.9.0"
sqlalchemy-utils = "^0.41.1"
psycopg2-binary = "^2.9.6"
flask-cors = "^3.0.10"
flask-migrate = "^4.0.4"
pytest-cov = "^4.1.0"
gunicorn = "^20.1.0"
pytz = "^2023.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
